Add more benchmarks and an initial interface
The interface seems to slow the uint64 version down by roughly a factor
of 3x, which is still 3x faster than the Big Int version. The Big Int
version is roughly the same speed with and without using the interface.

This makes sense, because most of the interface overhead is actually
just having to dereference a pointer, which is something the Big Int
version already had to do, which the uint64 version did not.

So the Big Int version basically gets a ~5ns / op slow down, which is
roughly the overhead of a function call that otherwise does nothing.